refactor: split telegram helpers
