+++ src/telegram/send.ts
@@ -1,4 +1,4 @@
-import { Chat, Message, Update } from "telegraf/types";
+import { Message } from "telegraf/types";
 import { useBot } from "../bot.ts";
 import { useConfig } from "../config.ts";
 import {
@@ -15,6 +15,8 @@ import {
   ForceReply,
 } from "telegraf/typings/core/types/typegram";
 import { log } from "../helpers.ts";
+import telegramifyMarkdown from "telegramify-markdown";
+import { splitBigMessage } from "../utils/text.ts";
 
 interface TelegramError extends Error {
   response?: {
@@ -22,37 +24,15 @@ interface TelegramError extends Error {
     description: string;
   };
 }
-type ForwardOrigin = {
+
+export type ForwardOrigin = {
   type: "user" | "hidden_user";
   sender_user?: User;
   sender_user_name?: string;
 };
-import telegramifyMarkdown from "telegramify-markdown";
 
 // let lastResponse: Message.TextMessage | undefined
 let forDelete: Message.TextMessage | undefined;
-
-export function splitBigMessage(text: string) {
-  const msgs: string[] = [];
-  const sizeLimit = 4096;
-  let msg = "";
-
-  for (const origLine of text.split("\n")) {
-    const line = origLine.trim();
-    if (!line) continue; // skip empty or whitespace-only lines
-    if (msg.length + line.length + 1 > sizeLimit) {
-      // +1 for the added '\n'
-      if (msg.trim()) msgs.push(msg);
-      msg = "";
-    }
-    msg += line + "\n";
-  }
-  if (msg.length > sizeLimit) {
-    msg = msg.slice(0, sizeLimit - 3) + "...";
-  }
-  if (msg.trim()) msgs.push(msg);
-  return msgs;
-}
 
 function sanitizeTelegramHtml(html: string): string {
   // Заменяем <p> и </p> на двойной перенос строки
@@ -68,7 +48,7 @@ function sanitizeTelegramHtml(html: string): string {
   return result.trim();
 }
 
-interface ExtraCtx {
+export interface ExtraCtx {
   noSendTelegram?: boolean;
   progressCallback?: (msg: string) => void;
 }
@@ -135,10 +115,7 @@ export async function sendTelegramMessage(
   if (params.parse_mode === "HTML") {
     processedText = sanitizeTelegramHtml(text);
   } else if (params.parse_mode === "MarkdownV2") {
-    // const replacedNewlines = text.replace(/\n/g, '#n');
-    // const ZWSP = '\u200B';
     processedText = telegramifyMarkdown(processedText, "keep");
-    // processedText = processedText.replace(/\\#n/g, `\n${ZWSP}`);
   } else if (params.parse_mode === "Markdown") {
     processedText = telegramifyMarkdown(text, "keep");
   }
@@ -162,13 +139,8 @@ export async function sendTelegramMessage(
           chatId: chat_id,
           logLevel: "warn",
         });
-        // Optionally: flag user in DB or take other action
         continue;
       }
-      // Fallback to failsafeParams for other errors
-      // Previous fallback code:
-      // const failsafeParams = { reply_markup: params.reply_markup };
-      // response = await useBot(chatConfig.bot_token).telegram.sendMessage(chat_id, msg, failsafeParams);
       const failsafeParams = {
         reply_markup: params.reply_markup as
           | InlineKeyboardMarkup
@@ -194,8 +166,8 @@ export async function sendTelegramMessage(
     if (response)
       setTimeout(async () => {
         await useBot(chatConfig?.bot_token).telegram.deleteMessage(
-          response.chat.id,
-          response.message_id,
+          response!.chat.id,
+          response!.message_id,
         );
       }, deleteAfter);
   }
@@ -213,11 +185,9 @@ export async function sendTelegramMessage(
     forDelete = response;
   }
 
-  // lastResponse = response
   return response;
 }
 
-// Check if the user is an admin
 export function isAdminUser(msg: Message.TextMessage): boolean {
   if (!msg.from?.username) return false;
   return (useConfig().adminUsers || []).includes(msg.from.username);
@@ -256,148 +226,6 @@ export function getFullName(msg: {
   }
 }
 
-function isAccessAllowed(chatConfig: ConfigChatType, ctxChat: Chat) {
-  const privateChat = ctxChat as Chat.PrivateChat;
-  const allowedUsers = [
-    ...(chatConfig.privateUsers ?? useConfig().privateUsers),
-    ...(useConfig().adminUsers ?? []),
-  ];
-  const username = privateChat.username || "without_username";
-  return allowedUsers.includes(username);
-}
-
-function getChatConfig(
-  ctxChat: Chat,
-  ctx: Context,
-): ConfigChatType | undefined {
-  // 1. by chat id
-  let chat =
-    useConfig().chats.find(
-      (c) => c.id == ctxChat?.id || c.ids?.includes(ctxChat?.id) || 0,
-    ) || ({} as ConfigChatType);
-
-  const defaultChat = useConfig().chats.find((c) => c.name === "default");
-
-  // 2. by bot_name
-  if (!chat.id) {
-    chat =
-      useConfig().chats.find((c) => c.bot_name === ctx.botInfo.username) ||
-      ({} as ConfigChatType);
-
-    // check access to private chat
-    if (chat.id && ctxChat?.type === "private") {
-      if (!isAccessAllowed(chat, ctxChat)) {
-        return;
-      }
-    }
-  }
-
-  if (!chat.id) {
-    // console.log("ctxChat:", ctxChat);
-    if (ctxChat?.type !== "private") {
-      const chatTitle = (ctxChat as Chat.TitleChat).title;
-      log({
-        msg: `This is ${ctxChat?.type} chat, not in whitelist: ${ctxChat.title}`,
-        chatId: ctxChat.id,
-        chatTitle,
-        logLevel: "warn",
-      });
-      return;
-    }
-
-    if (defaultChat) chat = defaultChat;
-
-    // 2. by username
-    if (ctxChat?.type === "private") {
-      // user chat, with username
-      const privateChat = ctxChat as Chat.PrivateChat;
-
-      // check access
-      if (!isAccessAllowed(chat, ctxChat)) {
-        return;
-      }
-
-      const userChat = useConfig().chats.find(
-        (c) => c.username === privateChat.username || "",
-      );
-      if (userChat) chat = userChat;
-    }
-  }
-
-  function mergeConfigParam<T extends Record<string, unknown>>(
-    name: keyof T,
-    from: Partial<T> | undefined,
-    to: T | undefined,
-  ) {
-    if (!from || !from[name] || !to) return;
-    to[name] = to[name]
-      ? ({
-          ...(from[name] as object),
-          ...(to[name] as object),
-        } as T[typeof name])
-      : from[name];
-  }
-
-  mergeConfigParam<{ completionParams: CompletionParamsType }>(
-    "completionParams",
-    useConfig() as Partial<{ completionParams: CompletionParamsType }>,
-    chat,
-  );
-
-  if (chat && defaultChat) {
-    chat = { ...defaultChat, ...chat };
-    mergeConfigParam<{ completionParams: CompletionParamsType }>(
-      "completionParams",
-      defaultChat,
-      chat,
-    );
-  }
-
-  return chat;
-}
-
-export function getActionUserMsg(ctx: Context): { user?: User; msg?: Message } {
-  // edited message
-  if (Object.prototype.hasOwnProperty.call(ctx, "update")) {
-    const updateQuery = ctx.update as Update.CallbackQueryUpdate;
-    const user = updateQuery.callback_query.from;
-    const msg = updateQuery.callback_query.message;
-    return { user, msg };
-  }
-  return {};
-}
-
-// return {chat, msg}
-export function getCtxChatMsg(ctx: Context): {
-  chat: ConfigChatType | undefined;
-  msg: Message.TextMessage | undefined;
-} {
-  let ctxChat: Chat | undefined;
-  let msg: Message.TextMessage | undefined;
-
-  // edited message
-  if (Object.prototype.hasOwnProperty.call(ctx, "update")) {
-    // console.log("ctx.update:", ctx.update);
-    const updateEdited = ctx.update as Update.EditedMessageUpdate; //{ edited_message: Message.TextMessage, chat: Chat };
-    const updateNew = ctx.update as Update.MessageUpdate;
-    msg = (updateEdited.edited_message ||
-      updateNew.message) as Message.TextMessage;
-    // console.log("msg:", msg);
-    ctxChat = msg?.chat;
-    // console.log('no message in ctx');
-    // return;
-  }
-
-  if (!ctxChat) {
-    console.log("no ctx chat detected");
-    return { chat: undefined, msg: undefined };
-  }
-
-  const chat = getChatConfig(ctxChat, ctx);
-
-  return { chat, msg };
-}
-
 export function getTelegramForwardedUser(
   msg: Message.TextMessage & { forward_origin?: ForwardOrigin },
   chatConfig: ConfigChatType,
@@ -416,7 +244,9 @@ export function getTelegramForwardedUser(
   const name =
     forwardOrigin.type === "hidden_user"
       ? forwardOrigin.sender_user_name
-      : `${forwardOrigin.sender_user?.first_name ?? ""} ${forwardOrigin.sender_user?.last_name ?? ""}`.trim();
+      : `${forwardOrigin.sender_user?.first_name ?? ""} ${
+          forwardOrigin.sender_user?.last_name ?? ""
+        }`.trim();
 
   return `${name}${username ? `, Telegram: @${username}` : ""}`;
 }
